Deal with (too) fast evaluation timings.
If the evaluation of an expression is sufficiently fast and the
overhead is overestimated sufficiently, the estimate of the true
evaluation time could be grossly exagerated.

USER VISIBLE CHANGE: The microbenchmark object (a matrix) may now
  contain 'NA' values if the measured time of evaluation is smaller than
  the estimated evaluation overhead.

RATIONAL: This is mainly a bugfix. In the past, this could lead to
  very large values in the result object.

WHAT TO CHANGE: You may need to adjust your analysis to deal with the
  'NA' values.