Multiple enhancements suggested by Hadley Wickham
microbenchmark() now retains the order of the expressions as they are
passed in. print.microbenchmark() has a new unit 'relative' and a new
argument 'order'. The 'relative' unit rescales all other times
relative to the time of the quickest w.r.t the median time. Using the
'order' argument, you can reorder the results according to any of the
columns displayed.

Apart from these major changes there are also several smaller bugfixes
in the handling of expressions and their representation as returned by
deparse().

Suggested-By: Hadley Wickham