Restore core timing routines, trim dependencies
Move ggplot2 and multcomp to "Suggests" and move core timing routines
from microbenchmarkCore back into one package.

autoplot.microbenchmark() is no longer officially registered as an S3
method, but the functionality is unchanged when microbenchmark is
loaded *and* attached. Both of the following calls to autoplot() will
work.

  library(microbenchmark)
  f <- function() NULL
  res <- microbenchmark(NULL, f(), times=1000L)
  boxplot(res)
  if (requireNamespace("ggplot2"))
    ggplot2::autoplot(res)
  if (require("ggplot2"))
    autoplot(res)

We only need to test for multcomp with requireNamespace(). We don't
need to anything if the package isn't installed, because we simply
ignore the multcomp output if there are any warnings or errors anyway.