Restore PKG_LIBS using Olaf's fix for R CMD check
The line below was rightly removed, because it uses the 'shell'
command, which isn't allowed by CRAN because it is a GNU extension.

PKG_LIBS = $(PKG_LIBS_$(shell uname -s))

But we still need to check if we're compiling on Linux, because
clock_gettime() was in librt until it was added to glibc in v2.17. Use
Olaf's solution from:

https://github.com/olafmersmann/microbenchmarkCore/blob/bef725cc2f579ae00bd54fd52e9f171ca6beff7e/src/Makevars

Ideally, we should check for clock_gettime() and only add librt to
PKG_LIBS if it can't be found, but this is likely good enough for most
platforms.