Initial attempt at autoconf
Instead of checking for specific operating systems, have configure
check for the desired functionality.

First check for high-resolution timing on macOS, so we can (eventually)
fall back to gettimeofday() if mach_absolute_time() is not available.

Next, check for clock_gettime() and a monotonic clockid_t. We check if
librt is needed (before clock_gettime() was in glibc).  The values for
clockid_t are checked in order of highest resolution to lowest.

Need to add support for gettimeofday().