Evaluate 'setup' and exprs in one environment
Create a new environment and ensure the setup and benchmark expressions
are all evaluated in that environment. This should allow the benchmark
expressions to find any objects defined in the 'setup' expression.
Thanks to Tim Keitt for providing an example demonstrating the issue.

Also add tests to check the 'setup' expression functionality whether
a check function is defined or not.

See #2.