Evaluate 'setup' and exprs in parent frame
Commit f628f5dda3539192c5268d26670d0467169d510f moved evaluation from
the parent frame to a new environment enclosed by the parent frame.
This broke existing functionality that allowed users to assign to a
(new) variable in the parent frame via the timed expression, which
caused at least 2 packages (RcppMsgPack and WeightedROC) to fail CRAN
checks.

Change the evaluation environment back to the parent frame (see #2).

Fixes #8.