Error in timing function if unit is invalid
Previously the error only occurred when printing the results. That
meant a long benchmark run would complete, but the results would not
print.

Add normalize_unit() to support variations of unit values. For example
"s", "secs", and "seconds" are all valid now.

Thanks to GitHub user @xrobin for the report.

Fixes #20.