Fix issue #35: Make y_max work in autoplot
The y_max parameter in autoplot.microbenchmark is currently ignored, because coord_cartesian and coord_flip do not work together. I removed coord_cartesian and moved ylim to coord_flip.

Now that y_max works, two issues appear in how it works: First, y_min cannot be 0 on a log axis, so I added a line to use the minimum value from object$ntime if log == TRUE. Second, the default value for y_max is derived from object$time, while the plot uses object$ntime, so I replaced the default with NULL and calculate the default value after calculating object$ntime. Note that the previous default for y_max also had no effect. This is why I did not keep the 1.05 factor, as it would change the behaviour compared to before.

Calling the function without y_max produces the same plots as before, but without the warning from coord_flip that it is replacing the coordinate system from coord_cartesian.